chore: fix all lints and formats
also adds these to CI to avoid regression
