+++ .github/workflows/ci.yaml
@@ -9,6 +9,42 @@ on:
       - v1.0
 
 jobs:
+  format:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Checkout Code
+      uses: actions/checkout@v3
+
+    - name: Set up Rust
+      uses: actions-rs/toolchain@v1
+      with:
+        toolchain: stable
+        profile: minimal
+        override: true
+
+    - name: check format
+      run: |
+        cargo fmt --check
+
+  lint:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Checkout Code
+      uses: actions/checkout@v3
+
+    - name: Set up Rust
+      uses: actions-rs/toolchain@v1
+      with:
+        toolchain: stable
+        profile: minimal
+        override: true
+
+    - name: check lint
+      run: |
+        cargo clippy
+
   build-and-test:
     runs-on: ubuntu-latest
