+++ crates/goose-cli/src/inputs.rs
