--- crates/goose-cli/src/inputs/mod.rs
@@ -1,1 +1,0 @@
-pub mod inputs;
