--- crates/goose-cli/src/profile/mod.rs
@@ -1,3 +1,0 @@
-pub mod profile;
-pub mod profile_handler;
-pub mod provider_helper;
