+++ crates/goose-cli/src/prompt/cliclack.rs
@@ -8,10 +8,7 @@ use bat::WrappingMode;
 use cliclack::{input, set_theme, spinner, Theme as CliclackTheme, ThemeState};
 use goose::models::message::{Message, MessageContent, ToolRequest, ToolResponse};
 
-use super::{
-    prompt::{Input, InputType, Prompt, Theme},
-    thinking::get_random_thinking_message,
-};
+use super::{thinking::get_random_thinking_message, Input, InputType, Prompt, Theme};
 
 pub struct CliclackPrompt {
     spinner: cliclack::ProgressBar,
