--- crates/goose-cli/src/prompt/mod.rs
@@ -1,4 +1,0 @@
-pub mod cliclack;
-pub mod prompt;
-pub mod rustyline;
-pub mod thinking;
