--- crates/goose-cli/src/session/mod.rs
@@ -1,2 +1,0 @@
-pub mod session;
-pub mod session_file;
